# Codigos/Back-End/docs/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API - Sistema de Controle de Fluxos Veiculares",
    "description": "API para gerenciamento de usuários, integradores, municípios, equipamentos e fluxos veiculares.",
    "version": "1.0.0",
    "contact": {
      "name": "DNIT",
      "url": "https://www.gov.br/dnit"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/",
      "description": "Servidor de desenvolvimento"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Equipamento": {
        "type": "object",
        "required": [
          "codigo",
          "faixa",
          "tipo",
          "ativo",
          "local",
          "marca",
          "modelo",
          "dataAfericao",
          "lacre",
          "integradorId",
          "municipioId"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "example": 1
          },
          "codigo": {
            "type": "string",
            "maxLength": 11,
            "example": "EQP001"
          },
          "faixa": {
            "type": "integer",
            "minimum": 1,
            "maximum": 127,
            "example": 1
          },
          "tipo": {
            "type": "string",
            "enum": ["CEV", "REV", "CEM"],
            "example": "CEV"
          },
          "ativo": {
            "type": "boolean",
            "example": true
          },
          "local": {
            "type": "string",
            "maxLength": 80,
            "example": "Rodovia BR-040, km 120"
          },
          "marca": {
            "type": "string",
            "maxLength": 40,
            "example": "XYZ Sensores"
          },
          "modelo": {
            "type": "string",
            "maxLength": 40,
            "example": "S1000"
          },
          "velocidadeLimite": {
            "type": "integer",
            "minimum": 0,
            "example": 80
          },
          "dataAfericao": {
            "type": "string",
            "format": "date",
            "example": "2025-01-15"
          },
          "lacre": {
            "type": "string",
            "maxLength": 20,
            "example": "L123456"
          },
          "dataRegistroInmetro": {
            "type": "string",
            "format": "date",
            "example": "2025-01-10"
          },
          "numeroInmetro": {
            "type": "string",
            "maxLength": 30,
            "example": "INM123456789"
          },
          "integradorId": {
            "type": "integer",
            "example": 1
          },
          "municipioId": {
            "type": "integer",
            "example": 1
          }
        }
      },
...
    },
    "parameters": {
      "idPath": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "integer"
        },
        "description": "ID do recurso"
      },
      "limitQuery": {
        "name": "limit",
        "in": "query",
        "schema": {
          "type": "integer",
          "default": 30,
          "maximum": 100
        },
        "description": "Limite de resultados retornados"
      }
    }
  },
  "paths": {
...
    "/equipamentos": {
      "get": {
        "tags": ["Equipamentos"],
        "summary": "Listar equipamentos",
        "parameters": [
          {
            "name": "codigo",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Filtrar por código do equipamento"
          },
          {
            "name": "faixa",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "Filtrar por faixa do equipamento"
          },
          {
            "name": "ativo",
            "in": "query",
            "schema": {
              "type": "boolean"
            },
            "description": "Filtrar por status ativo/inativo"
          },
          {
            "name": "integrador",
            "in": "query",
            "schema": {
              "type": "integer"
            },
            "description": "Filtrar por ID do integrador"
          },
          {
            "$ref": "#/components/parameters/limitQuery"
          },
          {
            "name": "dataInicio",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Data de início (YYYY-MM-DD)"
          },
          {
            "name": "dataFim",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "description": "Data de fim (YYYY-MM-DD)"
          },
          {
            "name": "horaInicio",
            "in": "query",
            "schema": {
              "type": "string",
              "pattern": "^([0-1][0-9]|2[0-3]):[0-5][0-9]$"
            },
            "description": "Hora de início (HH:MM)"
          },
          {
            "name": "horaFim",
            "in": "query",
            "schema": {
              "type": "string",
              "pattern": "^([0-1][0-9]|2[0-3]):[0-5][0-9]$"
            },
            "description": "Hora de fim (HH:MM)"
          },
          {
            "name": "placa",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Placa do veículo"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de equipamentos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Equipamento"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Equipamentos"],
        "summary": "Criar equipamento",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Equipamento"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Equipamento criado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Equipamento"
                }
              }
            }
          }
        }
      }
    },
    "/equipamentos/{id}": {
      "get": {
        "tags": ["Equipamentos"],
        "summary": "Obter equipamento por ID",
        "parameters": [
          {
            "$ref": "#/components/parameters/idPath"
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "number",
              "default": "30"
            },
            "description": "Limite de fluxos retornados"
          }
        ],
        "responses": {
          "200": {
            "description": "Equipamento encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Equipamento"
                }
              }
            }
          },
          "404": {
            "description": "Equipamento não encontrado"
          }
        }
      },
      "put": {
        "tags": ["Equipamentos"],
        "summary": "Atualizar equipamento",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "$ref": "#/components/parameters/idPath"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Equipamento"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Equipamento atualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Equipamento"
                }
              }
            }
          },
          "404": {
            "description": "Equipamento não encontrado"
          }
        }
      },
      "delete": {
        "tags": ["Equipamentos"],
        "summary": "Excluir equipamento",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "$ref": "#/components/parameters/idPath"
          }
        ],
        "responses": {
          "204": {
            "description": "Equipamento excluído com sucesso"
          },
          "404": {
            "description": "Equipamento não encontrado"
          }
        }
      }
    },
...
  }
}